check dir exists before clean up.

+++ src/maxdiffusion/tests/input_pipeline_interface_test.py
@@ -57,7 +57,8 @@ THIS_DIR = os.path.dirname(os.path.abspath(__file__))
 DATASET_DIR = str(HOME_DIR / ".cache" / "huggingface" / "datasets")
 
 def cleanup(output_dir):
-  shutil.rmtree(output_dir)
+  if os.path.isdir(output_dir):
+    shutil.rmtree(output_dir)
 
 class InputPipelineInterface(unittest.TestCase):
   """Test Unet sharding"""
